Fix accuracy of comma-grouped selectors
Use the jQuery `context` argument to limit the scope of the selector.

The rule was previously checked by prepending ".table-wrapper". When you
have a comma-grouped selector "A, B", this will result in:

    .table-wrapper A, B

It selects any "B", even ones outside of ".table-wrapper". This bug
shows up in the :last-of-type level, if you use the following selector:

    orange:last-of-type, apple:last-of-type

Even though this (rather unoptimized) selector should work, it
registered as a fail, because an innocent apple in the logo was also
selected. To that apple's credit, it joined the party and shook like
there's no tomorrow.