Show :first-child at bottom of stack
In the real world, the first item added to a stack would reside at the
bottom of the stack. However, before these changes, the first item added
to a stack (in the markup) was shown at the top of the stack. This made
level "Select the top orange" somewhat confusing.

These changes reverse the presentation, such that the first item added
to a stack is shown at the bottom of the stack.

The level "Select the small apple and the pickle" was rewritten to
account for this. Before these changes, the solution .small:last-child
would have been acceptable because the small orange at the top of the
stack was not considered to be the last item. Now that the item at the
top of the stack is considered to be the last item, the arrangement
would not have worked. Swapping the oranges also made little sense,
since a large orange would not normally rest on top of a small orange.
The foods were changed altogether, and a bento was added to keep things
visually interesting (otherwise it would have been nothing but plates,
oranges, and apples).